2y2s/Operating system and shells/Lab_2/requirements.txt
psutil==5.9.4
python-ptrace==0.9.8
